requirements.txt
PyYAML
tqdm
numpy
scipy
opencv-python
ffmpeg-python
nuscenes-devkit
